// src/types/model/study-task.model.types.ts
export enum eStudyTaskOptions {
	Chapter = 'chapter',
	Topic = 'topic',
}

export enum eStudyTaskOptionsChapterTags {
	Theory = 'theory',
	ShortNotes = 'shortNotes',
	PYQ_Mains = 'PYQ_Mains',
	PYQ_Advanced = 'PYQ_Advanced',
	DPP1 = 'DPP1',
	mindMap = 'mindMap',
	Module = 'Module',
	DPP2 = 'DPP2',
	Book = 'Book',
}

export enum eStudyTaskOptionsTopicTags {
	Theory = 'theory',
	InTextQuestions = 'inTextQuestions',
	InClassQuestions = 'inClassQuestions',
}